requirements.txt
httplib2
Flask
google-cloud
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-auth
